+++ spec/dummy/spec/lib/generators/apache_age/node_generator_spec.rb
@@ -28,6 +28,8 @@ RSpec.describe ApacheAge::NodeGenerator, type: :generator do
   NODEFILE
   }
 
+  after { described_class.start([node_name], {behavior: :revoke, destination_root:}) }
+
   context 'when using invoke' do
     let(:behavior) { :invoke }